Fixed overflow bug in AtomicLongTest (for increment tracking a long should be used instead of int)

+++ tests/tests-common/src/main/java/com/hazelcast/simulator/tests/concurrent/atomiclong/AtomicLongTest.java
@@ -117,7 +117,7 @@ public class AtomicLongTest {
 
     private class Worker extends AbstractWorker<Operation> {
 
-        private int increments;
+        private long increments;
 
         public Worker() {
             super(builder);
